+++ src/hitables/infinite_area_light.h
@@ -27,6 +27,8 @@ public:
   virtual bool HitP(const Ray &r, Float t_min, Float t_max, random_gen& rng) const;
   virtual bool HitP(const Ray &r, Float t_min, Float t_max, Sampler* sampler) const;
 
+  OpaqueShadowType ShadowType() const { return OpaqueShadowType::Light; }
+
   virtual bool bounding_box(Float t0, Float t1, aabb& box) const;
   virtual Float pdf_value(const point3f& o, const vec3f& v, random_gen& rng, Float time = 0);
   virtual Float pdf_value(const point3f& o, const vec3f& v, Sampler* sampler, Float time = 0);