+++ src/hitables/instance.h
@@ -23,6 +23,12 @@ public:
 
   vec3f random(const point3f& o, random_gen& rng, Float time = 0);
   vec3f random(const point3f& o, Sampler* sampler, Float time = 0);
+  OpaqueShadowType ShadowType() const {
+    return original_scene->ShadowType();
+  }
+  bool OpaqueHit(const Ray& r, Float t_min, Float t_max, random_gen& rng) const {
+    return original_scene->OpaqueHit((*WorldToObject)(r), t_min, t_max, rng);
+  }
   virtual bool bounding_box(Float t0, Float t1, aabb& box) const;
   
   virtual std::string GetName() const {
@@ -38,6 +44,7 @@ public:
   //Embedded scene
   hitable* original_scene;
   hitable_list* importance_sampled_objects;
+  LightPlacementMap light_placements;
 
 private:
   // Remap child boundary IDs into this placement's reserved range in its
@@ -46,4 +53,3 @@ private:
 };
 
 #endif
-