+++ src/hitables/plymesh.h
@@ -25,6 +25,13 @@ class plymesh : public hitable {
   virtual bool HitP(const Ray &r, Float t_min, Float t_max, random_gen& rng) const;
   virtual bool HitP(const Ray &r, Float t_min, Float t_max, Sampler* sampler) const;
 
+  OpaqueShadowType ShadowType() const {
+    return ply_mesh_bvh->ShadowType();
+  }
+  bool OpaqueHit(const Ray& r, Float t_min, Float t_max, random_gen& rng) const {
+    return ply_mesh_bvh->OpaqueHit(r, t_min, t_max, rng);
+  }
+
   virtual bool bounding_box(Float t0, Float t1, aabb& box) const;
   virtual std::string GetName() const {
     return(std::string("Plymesh"));